+++ Gruntfile.js
@@ -36,7 +36,6 @@ module.exports = function(grunt) {
   grunt.initConfig(config);
   grunt.config.merge(loadConfig('./tasks'));
 
-  grunt.loadNpmTasks('grunt-contrib-compass');
   grunt.loadNpmTasks('grunt-contrib-uglify');
   grunt.loadNpmTasks('grunt-contrib-concat');
   grunt.loadNpmTasks('grunt-gh-pages');
@@ -57,7 +56,6 @@ module.exports = function(grunt) {
 
   // Build the back office
   grunt.registerTask('build-back-office-client', [
-    'compass:back-office',
     'uglify:back-office',
     'concat:back-office-lib',
     'concat:back-office-js'