+++ compass.rb
@@ -1,0 +1,8 @@
+environment = :development
+css_dir = "./assets/be/css"
+sass_dir = "./app/client/admin/compass/sass"
+output_style = :nested
+sourcemap = true
+fonts_dir = "./node_modules/bootstrap-sass/assets/fonts/"
+http_fonts_path = "/bootstrap-sass/assets/fonts/"
+sass_options = {:precision => 10}